pyproject.toml
[project]
name = "promptsops"
version = "0.1.0"
description = "Local PromptsOps pipeline: evaluate, optimize, and version DSPy prompts with Ollama."
readme = "README.md"
authors = [
    { name = "Rohith Sundar Jonnalagadda", email = "rohithsundar.1392001@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix>=13.21.0",
    "datasets>=4.8.4",
    "dspy-ai>=3.1.3",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "openinference-instrumentation-dspy>=0.1.34",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.1",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
markers = [
    "integration: tests that require Ollama and local models",
    "slow: marks tests that call the LLM judge (deselect with '-m not slow')",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
files = ["src", "scripts", "tests"]
check_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true

[tool.coverage.run]
branch = true
source = ["promptsops"]

[tool.coverage.report]
fail_under = 60
show_missing = true
skip_empty = true
